backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.6",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "uvicorn>=0.40.0",
]
